Reimplement POST /api/v1/recordings as RequestHandlers
Also implement RequestHandler priorities, allowing multiple handlers to
be defined for the same method+route while also allowing predictable
handler ordering.
